requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.3.0
Flask==1.1.1
Flask-HTTPAuth==4.4.0 
gevent==1.4.0
greenlet==0.4.15
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==2.6.0
MarkupSafe==1.1.1
openapi-spec-validator==0.2.8
pathlib==1.0.1
Pillow==8.3.2
PyYAML==5.4
requests==2.22.0
six==1.12.0
swagger-ui-bundle==0.0.9
urllib3==1.26.5
Werkzeug==0.15.5
